Release 0.3.0

+++ src/ElectronNET.API/API/Entities/Margins.cs
@@ -1,33 +1,42 @@
 ﻿namespace ElectronNET.API.Entities;
 
 /// <summary>
-/// 
+/// Margins object used by webContents.print options and webContents.printToPDF.
 /// </summary>
+/// <remarks>Up-to-date with Electron API 39.2</remarks>
 public class Margins
 {
     /// <summary>
-    /// Can be `default`, `none`, `printableArea`, or `custom`. If `custom` is chosen,
+    /// Gets or sets the margin type. Can be `default`, `none`, `printableArea`, or `custom`. If `custom` is chosen,
     /// you will also need to specify `top`, `bottom`, `left`, and `right`.
     /// </summary>
     public string MarginType { get; set; }
 
     /// <summary>
-    /// The top margin of the printed web page, in pixels.
+    /// Gets or sets the top margin of the printed web page. Units depend on API:
+    /// - webContents.print: pixels
+    /// - webContents.printToPDF: inches
     /// </summary>
-    public int Top { get; set; }
+    public double Top { get; set; }
 
     /// <summary>
-    /// The bottom margin of the printed web page, in pixels.
+    /// Gets or sets the bottom margin of the printed web page. Units depend on API:
+    /// - webContents.print: pixels
+    /// - webContents.printToPDF: inches
     /// </summary>
-    public int Bottom { get; set; }
+    public double Bottom { get; set; }
 
     /// <summary>
-    /// The left margin of the printed web page, in pixels.
+    /// Gets or sets the left margin of the printed web page. Units depend on API:
+    /// - webContents.print: pixels
+    /// - webContents.printToPDF: inches
     /// </summary>
-    public int Left { get; set; }
+    public double Left { get; set; }
 
     /// <summary>
-    /// The right margin of the printed web page, in pixels.
+    /// Gets or sets the right margin of the printed web page. Units depend on API:
+    /// - webContents.print: pixels
+    /// - webContents.printToPDF: inches
     /// </summary>
-    public int Right { get; set; }
+    public double Right { get; set; }
 }
